Use double instead of float for precision improvement.
This change updates several float variables and function parameters to double to enhance precision across the codebase, particularly in camera and callback functionalities. This adjustment affects calculations related to time, position, and rendering to ensure more accurate and reliable results during program execution.

+++ camera.h
@@ -10,14 +10,14 @@ public:
     glm::vec3 target;
     glm::vec3 up;
 
-    float yaw, pitch, distance;
+    double yaw, pitch, distance;
 
     Camera();
     void updatePosition();
     glm::mat4 getViewMatrix();
-    void processMouseMovement(float xOffset, float yOffset);
-    void processMouseScroll(float yOffset);
-    void processKeyboard(GLFWwindow* window, float deltaTime);
+    void processMouseMovement(double xOffset, double yOffset);
+    void processMouseScroll(double yOffset);
+    void processKeyboard(GLFWwindow* window, double deltaTime);
 };
 
 #endif
